Cleaned up code and documentation
Col1, col2, col3, diag1, and diag 2 were variables not used anywhere and deleted. The endGame function was updated to a cleaner checking method. Not sure which method is faster, but it should not matter for this project. The printBoard function was merged into endGame as it was just a for each loop. Another call of endGame was added after playerInput to properly check if there was a win on the board. The board was updated to be filled with "-" instead of white space and the scratch game check was updated to work for this change. This change was made to have the board stay a consistent shape while playing in the terminal. 

The name of the file was changed so that it fit proper naming conventions and could be imported into other files. The GUI was moved to a separate file named GUI.